Park Transform functions have been added
The Park Transform block computes the Park transformation of two-phase orthogonal components (α, β) or multiplexed αβ0 components in a stationary αβ reference frame.

The block accepts the following inputs:

Either α-β axes components or multiplexed components αβ0 in the stationary reference frame. Use the Number of inputs parameter to use either two or three inputs.

Sine and cosine values of the corresponding angles of transformation.

When using two-input configuration, it outputs orthogonal direct (d) and quadrature (q) axis components in the rotating dq reference frame. When using three-input configuration, it outputs multiplexed components dq0.

For a balanced system, the zero component is equal to zero.

You can configure the block to align either the d- or the q-axis with the α-axis at time t = 0. bla bla bla. . .